refactor(test): clean shared test helpers and module wiring
